+++ src/lib/MessageWriter.ts
@@ -3,162 +3,196 @@ import * as D from './Decl';
 import { WsFrameEncoder } from './Encoder';
 import * as E from './Errors';
 import type * as $Net from 'node:net';
+import { Writable, WritableOptions } from 'node:stream';
 
-export class WsMessageWriteHelper {
+const EMPTY_BUFFER = Buffer.allocUnsafe(0);
 
-    public readonly encoder = new WsFrameEncoder();
+export class WsFrameWriter {
+
+    private readonly _encoder = new WsFrameEncoder();
 
     public maskKey: Buffer | boolean = false;
 
-    public constructor(
-        protected _socket: $Net.Socket | null,
-    ) {}
+    protected _socket: $Net.Socket | null;
 
-    public getMaskKey(): Buffer  | null {
+    public constructor(socket: $Net.Socket) {
 
-        return this.maskKey === true ? _.createRandomMaskKey() : (this.maskKey || null);
+        this._socket = socket.on('close', () => {
+
+            this._socket = null;
+        });
     }
 
-    public writePayload(data: Buffer | string | Array<Buffer | string>, maskKey: Buffer | null): boolean {
-
-        if (data instanceof Buffer) {
-
-            if (maskKey) {
-
-                this.encoder.mask(data, maskKey, 0);
-            }
-
-            return this._socket!.write(data);
-        }
-
-        if (Array.isArray(data)) {
-
-            if (maskKey) {
-
-                this.encoder.bulkMask(data, maskKey);
-            }
-
-            let ret: boolean = true;
-
-            for (const chunk of data) {
-
-                ret = this._socket!.write(chunk);
-            }
-
-            return ret;
-        }
-
-        if (maskKey) {
-
-            const buf = Buffer.from(data);
-
-            this.encoder.mask(buf, maskKey, 0);
-
-            return this._socket!.write(buf);
-        }
-
-        return this._socket!.write(data);
-    }
-}
-
-interface IBuffer {
-
-    content: string | Buffer;
-
-    maskKey: Buffer | null;
-
-    isFirst: boolean;
-}
-
-export class WsMessageWriter implements D.IMessageWriter {
-
-    public constructor(
-        public readonly opcode: D.EOpcode,
-        protected _socket: $Net.Socket | null,
-        protected readonly _helper: WsMessageWriteHelper
-    ) {}
-
-    protected _buffer: IBuffer | null = null;
-
-    protected _assertWritable(): void {
-
-        if (this._socket?.closed ?? true) {
-
-            throw new E.E_CONN_LOST();
-        }
+    /**
+     * Build and write a WebSocket frame
+     *
+     * @param opcode    The opcode of the frame
+     * @param isFin     Is this the final frame?
+     * @param payload   The payload of the frame
+     * @param callback  A callback to be called when the frame is written
+     *
+     * @returns Return false if the socket buffer is full, needs to wait for draining.
+     */
+    public writeArray(
+        opcode: D.EOpcode,
+        isFin: boolean,
+        payload: Array<Buffer | string>,
+        callback?: D.IErrorCallback
+    ): boolean {
 
         if (!this._socket?.writable) {
 
-            throw new E.E_CONN_READONLY();
+            if (callback) {
+                callback(new E.E_CONN_LOST());
+            }
+            else {
+                throw new E.E_CONN_LOST();
+            }
+
+            return false;
         }
-    }
 
-    public write(frame: Buffer | string): boolean {
+        let payloadLength = 0;
 
-        this._assertWritable();
+        for (const i of payload) {
 
-        const prev = this._buffer;
+            payloadLength += Buffer.byteLength(i);
+        }
 
-        const maskKey = this._helper.getMaskKey();
+        const maskKey = this.maskKey === true ? _.createRandomMaskKey() : (this.maskKey || null);
 
-        this._buffer = {
-            content: frame,
-            maskKey,
-            isFirst: !prev
-        };
+        if (maskKey) {
 
-        if (prev) {
+            this._encoder.bulkMask(payload, maskKey);
+        }
 
-            const l = Buffer.byteLength(prev.content);
+        const header = this._encoder.createHeader(opcode, payloadLength, isFin, maskKey);
 
-            const ret = this._socket!.write(this._helper.encoder.createHeader(
-                this.opcode,
-                l,
-                maskKey,
-                prev.isFirst ? 0 : 1,
-                false,
-            ));
+        if (!payloadLength) {
 
-            if (l) {
+            return this._socket.write(header, callback);
+        }
 
-                return this._helper.writePayload(prev.content, prev.maskKey) && ret;
+        let needDrain = this._socket.write(header);
+
+        for (let i = 0; i < payload.length; ++i) {
+
+            if (i === payload.length - 1) {
+
+                needDrain = this._socket.write(payload[i], callback);
+            }
+            else {
+
+                this._socket.write(payload[i]);
             }
         }
 
-        return false;
+        return needDrain;
     }
 
-    public end(frame?: string | Buffer): boolean {
+    /**
+     * Build and write a WebSocket frame
+     *
+     * @param opcode    The opcode of the frame
+     * @param isFin     Is this the final frame?
+     * @param payload   The payload of the frame
+     * @param callback  A callback to be called when the frame is written
+     *
+     * @returns Return false if the socket buffer is full, needs to wait for draining.
+     */
+    public write(
+        opcode: D.EOpcode,
+        isFin: boolean,
+        payload: Buffer | string,
+        callback?: D.IErrorCallback
+    ): boolean {
 
-        this._assertWritable();
+        if (!this._socket?.writable) {
 
-        if (frame) {
-
-            this.write(frame);
-        }
-
-        const prev = this._buffer;
-
-        this._buffer = null;
-
-        if (prev) {
-
-            const l = Buffer.byteLength(prev.content);
-
-            const ret = this._socket!.write(this._helper.encoder.createHeader(
-                this.opcode,
-                l,
-                prev.maskKey,
-                prev.isFirst ? 0 : 1,
-                true
-            ));
-
-            if (l) {
-
-                return this._helper.writePayload(prev.content, prev.maskKey) && ret;
+            if (callback) {
+                callback(new E.E_CONN_LOST());
             }
+            else {
+                throw new E.E_CONN_LOST();
+            }
+
+            return false;
         }
 
-        return false;
+        const payloadLength = Buffer.byteLength(payload);
+
+        const maskKey = this.maskKey === true ? _.createRandomMaskKey() : (this.maskKey || null);
+
+        if (maskKey) {
+
+            if (!(payload instanceof Buffer)) {
+
+                payload = Buffer.from(payload);
+            }
+
+            this._encoder.mask(payload, maskKey, 0);
+        }
+
+        const header = this._encoder.createHeader(opcode, payloadLength, isFin, maskKey);
+
+        if (!payloadLength) {
+
+            return this._socket.write(header, callback);
+        }
+
+        this._socket.write(header);
+
+        return this._socket.write(payload, callback);
+    }
+}
+
+export class WsMessageWriter extends Writable implements D.IMessageWriter {
+
+    private _sentBytes = 0;
+
+    public constructor(
+        public readonly opcode: D.EOpcode,
+        public readonly maxMessageSize: number,
+        protected readonly _writer: WsFrameWriter,
+        opts?: WritableOptions,
+    ) {
+
+        super(opts);
+    }
+
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    public override _write(
+        chunk: Buffer | string,
+        _encoding: BufferEncoding,
+        callback: D.IErrorCallback
+    ): void {
+
+        const len = Buffer.byteLength(chunk);
+
+        if (this._sentBytes + len > this.maxMessageSize) {
+
+            callback(new E.E_MESSAGE_TOO_LARGE());
+            return;
+        }
+
+        this._writer.write(
+            this._sentBytes === 0 ? this.opcode : D.EOpcode.CONTINUATION,
+            false,
+            chunk,
+            callback
+        );
+
+        this._sentBytes += len;
+    }
+
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    public override _final(callback: D.IErrorCallback): void {
+
+        this._writer.write(
+            this._sentBytes === 0 ? this.opcode : D.EOpcode.CONTINUATION,
+            true,
+            EMPTY_BUFFER,
+            callback
+        );
     }
 }